TST update warning message issued
