+++ sklearn/calibration.py
@@ -35,7 +35,7 @@ from .utils import (
 )
 
 from .utils.multiclass import check_classification_targets
-from .utils.fixes import delayed, Parallel
+from .utils.parallel import delayed, Parallel
 from .utils._param_validation import StrOptions, HasMethods, Hidden
 from .utils.validation import (
     _check_fit_params,
