+++ sklearn/inspection/_plot/partial_dependence.py
@@ -15,7 +15,7 @@ from ...utils import check_array
 from ...utils import check_matplotlib_support  # noqa
 from ...utils import check_random_state
 from ...utils import _safe_indexing
-from ...utils.fixes import delayed, Parallel
+from ...utils.parallel import delayed, Parallel
 from ...utils._encode import _unique
