+++ sklearn/linear_model/_coordinate_descent.py
@@ -31,7 +31,7 @@ from ..utils.validation import (
     column_or_1d,
 )
 from ..utils._readonly_array_wrapper import ReadonlyArrayWrapper
-from ..utils.fixes import delayed, Parallel
+from ..utils.parallel import delayed, Parallel
 
 # mypy error: Module 'sklearn.linear_model' has no attribute '_cd_fast'
 from . import _cd_fast as cd_fast  # type: ignore
