+++ sklearn/linear_model/_theil_sen.py
@@ -21,7 +21,7 @@ from ._base import LinearModel
 from ..base import RegressorMixin
 from ..utils import check_random_state
 from ..utils._param_validation import Interval
-from ..utils.fixes import delayed, Parallel
+from ..utils.parallel import delayed, Parallel
 from ..exceptions import ConvergenceWarning
 
 _EPSILON = np.finfo(np.double).eps
