+++ sklearn/metrics/pairwise.py
@@ -27,7 +27,7 @@ from ..utils import is_scalar_nan
 from ..utils.extmath import row_norms, safe_sparse_dot
 from ..preprocessing import normalize
 from ..utils._mask import _get_mask
-from ..utils.fixes import delayed, Parallel
+from ..utils.parallel import delayed, Parallel
 from ..utils.fixes import sp_version, parse_version
 
 from ._pairwise_distances_reduction import ArgKmin
