+++ sklearn/tests/test_config.py
@@ -4,7 +4,7 @@ from concurrent.futures import ThreadPoolExecutor
 import pytest
 
 from sklearn import get_config, set_config, config_context
-from sklearn.utils.fixes import delayed, Parallel
+from sklearn.utils.parallel import delayed, Parallel
 
 
 def test_config_context():
